tests/wpt/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrackList/length.html

<!doctype html>
<title>TextTrackList.length</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
setup(function(){
    window.video = document.createElement('video');
    video.addTextTrack('subtitles');
    window.track = document.createElement('track');
    video.appendChild(track);
    video.addTextTrack('subtitles');
});
test(function(){
    assert_equals(video.textTracks.length, 3);
});

</script>
